Refactor : Split admin.py in a package and multiple modules
